+++ test/test.js
@@ -40,7 +40,6 @@ function testTemplate(loader, template, options, testFn) {
 
   loader.call(new WebpackLoaderMock({
     query: options.query,
-    rootContext: options.rootContext || {},
     resolveStubs: resolveStubs,
     async: function (err, source) {
       if (err) {
@@ -122,101 +121,78 @@ describe('handlebars-loader', function () {
     });
   });
 
-  var partialResolverOverride = function partialResolverOverride(request, cb){
-    switch(request) {
+  it('should convert decorators into require statements', function (done) {
+    testTemplate(loader, './with-decorators.handlebars', {
+      stubs: {
+        './decorator': function(fn) { return fn; },
+        'decorator': function(fn) { return fn; }
+      },
+      data: TEST_TEMPLATE_DATA
+    }, function (err, output, require) {
+      assert.ok(output, 'generated output');
+      assert.ok(require.calledWith('decorator'),
+        'should have loaded decorator with module syntax');
+      assert.ok(require.calledWith('./decorator'),
+        'should have loaded decorator with relative syntax');
+      done();
+    });
+  });
+
+  function resolverOverride(resolve, request, type, cb){
+    switch (type) {
     case 'partial':
-      cb(null, 'test');
+      switch(request) {
+      case 'partial':
+        return cb(null, 'test');
+      case './partial':
+        return cb(null, './test');
+      }
       break;
-    case '$partial':
-      cb(null, './test');
+
+    case 'helper':
+      switch(request) {
+      case './image':
+        return cb(null, './helpers/image.js');
+      case './json':
+        return cb(null, './helpers2/json.js');
+      }
       break;
     }
-  };
 
-  var testPartialResolver = function testPartialResolver(expectation, options, config){
-    it(expectation, function (done) {
-      testTemplate(loader, './with-partials.handlebars', {
-        stubs: {
-          './test': require('./partial.handlebars'),
-          'test': require('./partial.handlebars')
-        },
-        query: {
-          config: config
-        },
-        rootContext: options,
-        data: TEST_TEMPLATE_DATA
-      }, function (err, output, require) {
-        assert.ok(output, 'generated output');
-        assert.ok(require.calledWith('test'),
-          'should have loaded partial with module syntax');
-        assert.ok(require.calledWith('./test'),
-          'should have loaded partial with relative syntax');
-        done();
-      });
+    return cb();
+  }
+
+  it('should use the resolver for partials if specified', function (done) {
+    testTemplate(loader, './with-partials.handlebars', {
+      query: {
+        resolver: resolverOverride
+      },
+      stubs: {
+        './test': require('./partial.handlebars'),
+        'test': require('./partial.handlebars')
+      },
+      data: TEST_TEMPLATE_DATA
+    }, function (err, output, require) {
+      assert.ok(output, 'generated output');
+      assert.ok(require.calledWith('test'),
+        'should have loaded partial with module syntax');
+      assert.ok(require.calledWith('./test'),
+        'should have loaded partial with relative syntax');
+      done();
     });
-  };
+  });
 
-  testPartialResolver(
-    'should use the partialResolver if specified', {
-      handlebarsLoader: {
-        partialResolver: partialResolverOverride
-      }
-    }
-  );
-
-  testPartialResolver(
-    'honors the config query option when finding the partialResolver', {
-      handlebarsLoaderOverride: {
-        partialResolver: partialResolverOverride
-      }
-    },
-    'handlebarsLoaderOverride'
-  );
-
-  var helperResolverOverride = function helperResolverOverride(request, cb){
-    switch(request) {
-    case './image':
-      cb(null, './helpers/image.js');
-      break;
-    case './json':
-      cb(null, './helpers2/json.js');
-      break;
-    default:
-      cb();
-    }
-  };
-
-  var testHelperResolver = function testHelperResolver(expectation, options, config){
-    it(expectation, function (done) {
-      testTemplate(loader, './with-dir-helpers-multiple.handlebars', {
-        query: {
-          config: config
-        },
-        rootContext: options,
-        data: TEST_TEMPLATE_DATA
-      }, function (err, output, require) {
-        assert.ok(output, 'generated output');
-        done();
-      });
+  it('should use the resolver for helpers if specified', function (done) {
+    testTemplate(loader, './with-dir-helpers-multiple.handlebars', {
+      query: {
+        resolver: resolverOverride
+      },
+      data: TEST_TEMPLATE_DATA
+    }, function (err, output, require) {
+      assert.ok(output, 'generated output');
+      done();
     });
-  };
-
-  testHelperResolver(
-    'should use the helperResolver if specified', {
-      handlebarsLoader: {
-        helperResolver: helperResolverOverride
-      }
-    }
-  );
-
-  testHelperResolver(
-    'honors the config query option when finding the helperResolver', {
-      handlebarsLoaderOverride: {
-        helperResolver: helperResolverOverride
-      }
-    },
-    'handlebarsLoaderOverride'
-  );
+  });
 
   it('allows specifying additional helper search directory', function (done) {
     testTemplate(loader, './with-dir-helpers.handlebars', {
@@ -336,7 +312,8 @@ describe('handlebars-loader', function () {
       query: '?runtime=handlebars/runtime.js', // runtime actually gets required() as part of version check, so we specify real path to runtime but specify the extension so we know loader is using our custom version.
       stubs: {
         'handlebars/runtime.js': {
-          default: handlebarsAPI
+          default: handlebarsAPI,
+          __esModule: true
         }
       }
     }, function (err, output, require) {
@@ -367,7 +344,7 @@ describe('handlebars-loader', function () {
 
     async.series([
       testWithHandlebarsAPI({ template: templateStub }), // CommonJS style
-      testWithHandlebarsAPI({ default: { template: templateStub } }) // ES6 style
+      testWithHandlebarsAPI({ default: { template: templateStub }, __esModule: true }) // ES6 style
     ], function (err, results) {
       assert.ok(!err, 'no errors');
       assert.ok(results.filter(Boolean).length === 2, 'generated output');
@@ -410,7 +387,10 @@ describe('handlebars-loader', function () {
 
     // Need to set up a stubbed handlebars runtime that has our known helper loaded in
     var Handlebars = require('handlebars/runtime').default.create();
-    stubs[runtimePath] = Handlebars;
+    stubs[runtimePath] = {
+      default: Handlebars,
+      __esModule: true
+    };
     Handlebars.registerHelper('someKnownHelper', function () {
       return 'some known helper';
     });
@@ -526,13 +506,20 @@ describe('handlebars-loader', function () {
 
     // Need to set up a stubbed handlebars runtime that has our known helper loaded in
     var Handlebars = require('handlebars/runtime').default.create();
-    stubs[runtimePath] = Handlebars;
+    stubs[runtimePath] = {
+      default: Handlebars,
+      __esModule: true
+    };
     Handlebars.registerHelper('someKnownHelper', function () {
       return 'some known helper';
     });
 
     testTemplate(loader, './with-known-helpers.handlebars', {
-      query: '?ignoreHelpers',
+      query: {
+        resolver(resolve, request, type, callback) {
+          callback();
+        }
+      },
       stubs: stubs
     }, function (err, output, require) {
       assert.ok(output, 'generated output');